airodump-ng: retain fractional advertised rates

+++ include/aircrack-ng/support/station.h
@@ -120,7 +120,7 @@ struct AP_info
 	struct n_channel_info n_channel; /* 802.11n channel info     */
 	struct ac_channel_info ac_channel; /* 802.11ac channel info    */
 	struct ax_channel_info ax_channel; /* 802.11ax channel info    */
-	int max_speed; /* AP maximum speed in Mb/s */
+	float max_speed; /* AP maximum speed in Mb/s */
 	int bss_load_station_count; /* Stations from BSS Load IE */
 	int avg_power; /* averaged signal power    */
 	int best_power; /* best signal power    */
